chore: Reset input fields in the sidebar
This commit adds a new method, `reset_input_fields`, to the `TunisianRamiGame` class in the `streamlit_app.py` file. This method is responsible for resetting the input fields in the sidebar by setting the values to 0. It improves the user experience by allowing the players to easily reset their input values when needed.